Add libcompat integrating clocks
First, this patch changes gettimeofday and settimeofday function call by
clock_gettime and clock_settime using the CLOCK_REALTIME id. The reason
for that is for POSIX compliance and portability.

Thinking ahead for a possible Windows port, this patch also introduce a
basic compatibility layer where generic function can be used which
points to the right OS code. A libcompat is available which for now only
provide an interface for clocks.

Refer to src/compat/clock.h for the possible calls and options. The
tlsdate-helper was refactored to use this libcompat and new clock calls.

Also silent compilation output so warnings and errors can easily be seen
during compile time.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>