rwx permissions in save_disk_timestamp and write_time https://github.com/ioerror/tlsdate/issues/66
The function save_disk_timestamp creates a file with user read, write and execute permissions. The execute permission is unnecessary because it is guaranteed to be a file with the time_t in it.

The proper permission is S_IRUSR | S_IWUSR.

Similarly the function write_time in src/tlsdated-unittest.c creates a file with user read, write and execute permissions.