after a certain point time_map should not be zero
While porting tlsdate to Debian GNU/Hurd 7.0 I noticed that we would return
zero as a valid server time. This happens as a result of some strange mmap and
memcpy interactions on the Hurd. This could happen if the server was a false ticker
or as a result of some other strange flukes, so we should never accept it and
if time_map is zero, we should stop trying to use this server. We die when we
encounter zero and it is probably the case that a smart administrator who
wanted to mess with tlsdate users would return 32bits of zeros. We consider
this useful as it is wrong and detectable as such.