tlsdated: remove rmrf()
This function, while handy and presently used safely, could become dangerous if
someone later adds a call to it that passes in attacker-controlled input.
Therefore, hardcode calls to unlink(2) and rmdir(2) for the files tlsdated
presently creates. This has the pleasant (?) side-effect of breaking unit-tests
whenever tlsdated creates files we don't explicitly list on disk.

BUG=None
TEST=unit

Signed-off-by: Elly Fong-Jones <ellyjones@chromium.org>