tlsdate-helper: connect to all IP addresses for the host in parallel.
For example, if a given host has both IPv4 and IPv6 addresses, and your
client device thinks it has a valid IPv6 route but that route doesn't work,
tlsdate-helper could delay for a very long time as it failed on the IPv6
addresses one by one.  This way it will try the IPv4 addresses at the same
time, and just use the first socket that connects successfully, so there
is no delay.